backend/requirements.txt
django-cors-headers==4.4.0
djangorestframework==3.15.2
mysqlclient==2.2.4
pip-chill==1.0.3
pymysql==1.1.1
python-decouple==3.8
pillow==10.3.0
PyJWT==2.3.0